requirements.txt
pillow
torch
torchvision
opencv-python
matplotlib
pandas
ultralytics
plotly
keras
